requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
dotenv==0.9.9
idna==3.11
kafka-python==2.2.16
loguru==0.7.3
pillow==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pythorhead==0.34.3
requests==2.32.5
semver==3.0.4
six==1.17.0
tenacity==9.1.2
urllib3==2.5.0
